Initial working implementation for pathfinding in the first region ...
Separated the RomMap from the Tile class and made TileGraph to store the
TileNode data in a list. Added path finding capabilities using a
flood fill / breadth first search to discover all key areas and then
reconstruct the path back to all reachable key areas. No logic yet for
acquring rewards to traverse past the first region / palace of the map.
Added visualization capabilities using moviepy as the main output for
the moment. Output consists of the flood fill concatenated with the path
finding reconstruction. Re-worked the directory structure for typical
package configuration and driver class in the root of the project directory
as the main method of running things