Interim work toward topological graph formation
Moved the inner test module into the main module and started to
move the tile module into a configuration manager for each type
of configuration file extracted from the main rom. Turned the main
file into the start of the test suite. Transitioning the approach
to forming a graph that a topological sorting algorithm can
operate on in comparison to attempting to solve each section in
chunks as this approach feels better suited for the actual problem.