Graph formation feature (#1)
* Interim work toward topological graph formation

Moved the inner test module into the main module and started to
move the tile module into a configuration manager for each type
of configuration file extracted from the main rom. Turned the main
file into the start of the test suite. Transitioning the approach
to forming a graph that a topological sorting algorithm can
operate on in comparison to attempting to solve each section in
chunks as this approach feels better suited for the actual problem.

* Add initial topological graph formation method

* Complete the topological graph formation algorithm

Using the imported map data, this algorithm creates a graph using
the associated costs and rewards for linking tile together.
Then it processes the bottlenecks for traversal in stages and adds
these bottlenecks to the graph

* Current working status for map -> graph processing

* Remove the RomMap data structure ...

This data is specific to zelda2 and I wish to incorporate it
as a test rather than in the library itself to make things more
general

* Eliminate the dedicated logging file ...

Handle internally with data structures themselves

* Handle analysis / measurement external to library

* Handle items external to library

* Eliminate the half baked path handling files

* Move all the map data structures into specific mod

* Centralizing the map data structures in one place

* Replacing the map_data with conftest for setup ...

Rather than having a dedicated "map_data" class hardcoded
to the zelda2 map test example, we have the configuration for
building that map in the conftest as a fixture that
we then use for flushing out the base data for our test
methods

* Update the requirements for new pytest plugins

* Simplify test structure / module structure

* Add seperate processing for location data from map

* Normalize module permissions to 644

* Separate the explore_region method from pathprocessor

* Remove the explore_region method duplicate

* Refactor the form_tile_graph into smaller parts

* Update the conftest file to include the zelda 2 example

* Update the pyproject build file

* Interim commit for full graph generation refactor

* Cleanup tests + tilelocations test implementation

* Update the tilemap tests

* In progress development for graph bottlenecks

* Correct map generation test

* Remove file instances of path processing

* Remove old configuration logic

* Complete restructure of library + test layout

* Update Makefile and configuration files

* Initial implementation for topological sort ...

Likely need to approach the problem from the inverse to properly form
the correct topologically pathing requirements

* Correct type checking to avoid circular import

* Separate notion of bottleneck from graph sorting

* Update the configuration file indentation

* Update the README to provide minimum examples