Change fixtures to allow communal data storage ...
Move zelda2 processing back to main test directory. Changes
to the fixtures within the conftest allow for one communal area for
storing and processing the test data to no longer require separate
folders

Also initial elimination of the rom and environement requirements
for running the zelda2 tests. Instead provide a tool for generating
the map data and simply store the map data as plaintext within the
new communal data storage on the repository for simplification